user-service/feature/user-register: refactor project structure
