+++ user-service/internal/routers/user.router.go
@@ -2,7 +2,7 @@ package routers
 
 import (
 	"github.com/gin-gonic/gin"
-	"music-streaming-microservices/user-service/wire"
+	"music-streaming-microservices/user-service/pkg/wire"
 )
 
 type UserRouter struct {
