+++ user-service/pkg/wire/user.wire.go
